pyproject.toml
[tool.poetry]
name = "python-powerpoint-dashboard"
version = "0.1.0"
description = ""
authors = ["James Neale <42097394+james9226@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.30.0"
python-pptx = "^0.6.23"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
